Refactored toolbar out of the activities
